+++ lib/jsx-runtime.d.ts
@@ -1,4 +1,4 @@
 /* eslint-disable @typescript-eslint/triple-slash-reference */
 
 /// <reference path="./modules/JSX/jsx.d.ts" />
-export {jsx, jsxs, Fragment} from './modules/JSX/runtime';
+export {jsx, jsxs, jsxDEV, Fragment} from './modules/JSX/runtime';